requirements.txt
yfinance
pandas
scikit-learn
statsmodels
numpy
scipy
google-genai
asyncio
